% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropRawImages.R
\name{cropRawImages}
\alias{cropRawImages}
\title{Crop raw scanner images}
\usage{
cropRawImages(
  exp.design,
  crop.coords = "crop.coords",
  start.from = NA,
  patch.up = NA,
  path.to.ij.jar = "/Applications/ImageJ.app/Contents/Java/ij.jar"
)
}
\arguments{
\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{crop.coords}{Column of the design table that contains cropping coordinates. Defaults to "crop.coords"}

\item{start.from}{If you started this process earlier and are continuing mid-way, specify the scan.prefix to start with.}

\item{patch.up}{If you did this process already but want to change the crop for one plate, specify the scan.prefix of the plate.}

\item{path.to.ij.jar}{Location of imageJ jar file. Defaults to "/Applications/ImageJ.app/Contents/Java/ij.jar".}
}
\description{
This function cleans up the metadata table, and allows you to specify roles for columns.
}
\examples{
cropRawImages()
}
