% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCrops.R
\name{drawCrops}
\alias{drawCrops}
\title{Define crop coordinates}
\usage{
drawCrops(
  exp.design,
  start.from = NA,
  patch.up = NA,
  save.to = "./tmp_meta_with_crop_coords.csv",
  path.to.ij.jar = "/Applications/ImageJ.app/Contents/Java/ij.jar"
)
}
\arguments{
\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{start.from}{If you started this process earlier and are continuing mid-way, specify the scan.prefix to start with.}

\item{patch.up}{If you did this process already but want to change the crop for one plate, specify the scan.prefix of the plate.}

\item{save.to}{Location where crop coordinates will be saved.}

\item{path.to.ij.jar}{Location of imageJ jar file. Defaults to "/Applications/ImageJ.app/Contents/Java/ij.jar".}
}
\description{
This function opens scanner images in imageJ, prompts you to draw the cropping circle, and saves the coordinates.
}
\examples{
drawCrops()
}
