% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{growthCurves}
\alias{growthCurves}
\title{Convert pixel intensity tables to growth curves}
\usage{
growthCurves(
  exp.design,
  start.from = NA,
  patch.up = NA,
  remove.minimum = F,
  save.to = "./tmp_standard_deviations.csv",
  skip.filtereds = TRUE
)
}
\arguments{
\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{remove.minimum}{For each pixel intensity, option to subtract the smallest frequency in the series from all images in the series. This should reduce background and improve contrast. Defaults to FALSE.}

\item{save.to}{Location where standard deviation file will be saved. Defaults to "./tmp_standard_deviations.csv".}
}
\description{
This function opens tables with pixel intensity information, and calculates standard deviation, to represent worm density.
}
\examples{
growthCurves()
}
