% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelTables.R
\name{pixelTables}
\alias{pixelTables}
\title{Convert images to pixel intensity frequency tables}
\usage{
pixelTables(
  exp.design,
  start.from = NA,
  patch.up = NA,
  path.to.ij.jar = "/Applications/ImageJ.app/Contents/Java/ij.jar"
)
}
\arguments{
\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{start.from}{If you started this process earlier and are continuing mid-way, specify the scan.prefix to start with.}

\item{patch.up}{If you did this process already but want to change the crop for one plate, specify the scan.prefix of the plate.}

\item{path.to.ij.jar}{Location of imageJ jar file. Defaults to "/Applications/ImageJ.app/Contents/Java/ij.jar".}
}
\description{
This function opens (cropped) scanner images in imageJ and calculates how many pixels there are of each pixel intensity (from white to black). Returns a table.
}
\examples{
pixelTables()
}
