% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToStarve.R
\name{timeToStarve}
\alias{timeToStarve}
\title{Takes a growth curve and finds the peak, calculating hours from feeding to starvation}
\usage{
timeToStarve(
  sd.table,
  exp.design,
  inspect.each = F,
  save.to = "tmp_meta_with_time_to_starve.csv"
)
}
\arguments{
\item{sd.table}{A table with one row per image captured, with standard deviation of pixel intensity already calculated.}

\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{inspect.each}{Allows you to inspect each trace by eye to make sure the min and max points of the growth curve are correctly IDed, and edit the "trim_before" and "trim_after" specifications, if not.}

\item{save.to}{Location to save the new meta file, now with peak and valley values, and edited trim_before and trim_afters. Defaults to "./tmp_meta_with_time_to_starve.csv"}
}
\description{
This function accepts a table with a "sd" value (standard deviation of pixel intensity) for each image captured in a time course. It finds the peak and calculates the time of starvation.
}
\examples{
timeToStarve()
}
