% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.R
\name{tm}
\alias{tm}
\title{Covert a yy-mm-dd_HH-MM character into a time object}
\usage{
tm(time.in)
}
\arguments{
\item{time.in}{Character representing time, in the format "yy-mm-dd_HH-MM".}
}
\description{
This function takes a character with a specific format (yy-mm-dd_HH-MM), and converts it to a time format.
}
\examples{
tm()
}
