% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ij_script.R
\name{write_ij_script}
\alias{write_ij_script}
\title{Create ImageJ macro scripts}
\usage{
write_ij_script(which.script)
}
\arguments{
\item{which.script}{Which script would you like to create? Options: "pix_int", "draw_crops", crop_images", or "all"}
}
\description{
This function generates the macro scripts needed for imageJ functions. It writes them to a findable location.
}
