% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatios.R
\name{plotRatios}
\alias{plotRatios}
\title{Takes a growth curve table and meta_with_minmax table and outputs an assortment of plots}
\usage{
plotRatios(
  sd.table,
  exp.design,
  save.to.folder = "./tmp_plots",
  plot.by.sens.score = TRUE,
  plot.ctrl.by.exp = TRUE,
  wide.view = FALSE,
  p.width = 7,
  p.height = 5
)
}
\arguments{
\item{sd.table}{A table with one row per image captured, with standard deviation of pixel intensity already calculated.}

\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{save.to.folder}{The folder where images will be saved. Defaults to "./tmp_plots".}

\item{plot.by.sens.score}{Plot ratios by sensitivity. Defaults to TRUE.}

\item{plot.ctrl.by.exp}{Plot ctrl values on one axis and experimental values on the other. Defaults to TRUE.}

\item{wide.view}{For ctrl-by-exp plot, make the x and y axes limits match, and add a 1:1 reference line. Makes square dotplot. Defaults to FALSE.}

\item{p.width}{Width of ratio plots. Defaults to 7.}

\item{p.height}{Height of ratio plots. Defaults to 5.}

\item{plot.unfaceted.tts}{Plot time to starvation, all on one plot, with samples ordered by sensitivity score. Defaults to TRUE.}
}
\description{
Takes a growth curve table and meta_with_minmax table and outputs an assortment of plots
}
\examples{
plotRatios()
}
