% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatios.R
\name{plotRatios}
\alias{plotRatios}
\title{Takes a growth curve table and meta_with_minmax table and outputs an assortment of plots}
\usage{
plotRatios(
  sd.table,
  exp.design,
  save.to.folder = "./tmp_plots",
  ratio.trait = "colorer",
  plot.by.sens.score = TRUE,
  plot.ctrl.by.exp = TRUE,
  wide.view = TRUE,
  radiate = NULL,
  plot.unfaceted.tts = TRUE,
  pval.reds = FALSE,
  pairwise.sens.scores = TRUE,
  p.width = 5,
  p.height = 4,
  p.suffix = ""
)
}
\arguments{
\item{sd.table}{A table with one row per image captured, with standard deviation of pixel intensity already calculated.}

\item{exp.design}{A table with one row per culture/plate/petri dish. Each column adds info about how to group, facet, color, etc the data points.}

\item{save.to.folder}{The folder where images will be saved. Defaults to "./tmp_plots".}

\item{ratio.trait}{Name of the column which describes the variable used in sensitivity score ratios. Defaults to "colorer", which might be dose, depending on how you set up your meta.}

\item{plot.by.sens.score}{Plot ratios by sensitivity. Defaults to TRUE.}

\item{plot.ctrl.by.exp}{Plot ctrl values on one axis and experimental values on the other. Defaults to TRUE.}

\item{wide.view}{For ctrl-by-exp plot, make the x and y axes limits match, and add a 1:1 reference line. Makes square dotplot. Defaults to TRUE}

\item{radiate}{For wide ctrl-by-exp plot, draw a 1:x and x:1 reference lines, in addition to the 1:1. Defaults to NULL.}

\item{plot.unfaceted.tts}{Plot time to starvation, all on one plot, with samples ordered by sensitivity score. Defaults to TRUE.}

\item{pairwise.sens.scores}{Whether to make a grid plot colored by p-values showing statistical significance of different strains responses to the treatment. Defaults to TRUE.}

\item{p.width}{Width of ratio plots. Defaults to 7.}

\item{p.height}{Height of ratio plots. Defaults to 5.}

\item{p.suffix}{If running this multiple ways, include a suffix for the exported files, to distinguish them from each other. Defaults to "" (nothing).}

\item{pval.red}{Whether to color code ratios by statistical significance. Defaults to F - all black dots.}
}
\description{
Takes a growth curve table and meta_with_minmax table and outputs an assortment of plots
}
\examples{
plotRatios()
}
